requirements.txt
# LangChain & LangGraph
langchain>=1.0.0,<2.0.0
langgraph>=1.0.0,<2.0.0
langchain-openai>=1.0.0,<2.0.0
langchain-text-splitters>=1.0.0,<2.0.0
langchain-community>=0.4.0
langchain-anthropic>=1.0.0,<2.0.0
langchain-google-genai>=2.0.0
langsmith>=0.2.0

# 기타 의존성
beautifulsoup4>=4.12.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# PDF 처리
pypdf>=3.0.0
pillow>=10.0.0

# 웹 프레임워크
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# 데이터베이스 드라이버
psycopg2-binary>=2.9.0,<3.0.0

# 체크포인터 (선택사항 - DB 기반 체크포인터 사용 시)
# 최신 버전: 3.0.4 (psycopg 3.x 사용)
langgraph-checkpoint-postgres>=1.0.0

